requirements.txt
pandas
numpy
ktrain
torch
pytorch_lightning
Flask
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
-e .